refactor(form): inline shared helpers into form-interactive.tsx
Collapse form-shared.tsx back into form-interactive.tsx so the recipe
lives in a single file. Helpers (schema, copy, layout shell, hooks,
FieldRow components) stay above the recipes; nothing else changes.

- src/app/demo/[name]/ui/form-shared.tsx: deleted
- registry.json: drop the form-shared file entry from the form example
- src/app/(registry)/registry/[name]/page.tsx: drop the \${componentName}-shared
  lookup added when form-shared existed
